pyproject.toml
[project]
name = "prompt_agent"
version = "0.1.0"
description = "prompt-agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "crewai[tools]==1.9.3",
    "email-validator>=2.2.0",
    "fastapi>=0.135.1",
    "fastapi-sso>=0.18.0",
    "langchain-groq>=1.1.2",
    "litellm>=1.75.3",
    "opik>=1.8.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.0",
]

[project.scripts]
prompt_agent = "prompt_agent.main:run"
run_crew = "prompt_agent.main:run"
train = "prompt_agent.main:train"
replay = "prompt_agent.main:replay"
test = "prompt_agent.main:test"
run_with_trigger = "prompt_agent.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
